pyproject.toml
[build-system]
requires = [
    "setuptools>=65.0.0",
    "wheel>=0.40.0"
]
build-backend = "setuptools.build_meta"

[project]
name = "ltm"
version = "0.1.0"
description = "lightweight Telegram notification lib"
authors = [
    { name = "DiorditsPV", email = "pavel.diordits@yandex.ru" }
]
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "requests>=2.32.3",
    "python-dotenv>=0.19.0"
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
ltm = ["py.typed"] 